Create Rating_Product_and_Sorting_Reviews_in_Amazon.py
One of the most important problems in e-commerce is selling products
is the correct calculation of the points awarded. This
the solution to the problem is more customers for the e-commerce site
satisfaction, product prominence and purchase for sellers
means a seamless shopping experience for the recipients. Another
the problem is that the comments given to the products
as a ranking. Misleading comments stand out
will directly affect the sale of the product, so it is important to both
It will cause both financial loss and customer loss. These 2 basic
e-commerce site and sellers in solving the problem
and customers are able to complete the purchase journey seamlessly.
will complete it.